refactor: right-size wire field types for replication_factor, cache_extents, extent_growth_factor
Narrow wire protocol fields to their natural sizes:
- replication_factor: u16 → u8 (max RF=5, saves 1 byte/frame)
- cache_extents: u32 → u16 (max 65535 extents, saves 2 bytes/frame)
- extent_growth_factor: u32 → u8 (max 255, saves 3 bytes/frame)

Update all encode/decode, size calculations, Rust types, SQL schema
(SMALLINT→TINYINT UNSIGNED, INT→SMALLINT UNSIGNED/TINYINT UNSIGNED),
metadata reads/binds, config defaults, and design.md wire layouts.

+++ benches/pipeline_append.rs
@@ -59,8 +59,8 @@ const REPLICATION_FACTOR: u16 = 2;
 const MIN_EXTENT_CAPACITY: u32 = 8 * 1024 * 1024; // 8 MiB
 const MAX_EXTENT_CAPACITY: u32 = 256 * 1024 * 1024; // 256 MiB
 const PIPELINE_DEPTH: usize = 16; // max in-flight appends per sender
-const CACHE_EXTENTS: u32 = 4;
-const EXTENT_GROWTH_FACTOR: u32 = 8; // 8x growth: 8→64→256 MiB (2 transitions vs 5 with 2x)
+const CACHE_EXTENTS: u16 = 4;
+const EXTENT_GROWTH_FACTOR: u8 = 8; // 8x growth: 8→64→256 MiB (2 transitions vs 5 with 2x)
 
 // -- Shared counters ----------------------------------------------------------
